Add watchdog timer
This firmware is getting quite complicated. There's no way I can
guarantee it won't crash at this point. The watchdog should reboot if it
locks up for 5 seconds. This could feasibly happen if the API is really
slow, but I can add timeouts to the web requests to hopefully avoid
that.